// packages/user-connectrpc-server/src/generated/chat3_user_pb.d.ts
export declare type SubscribeUpdatesRequest = Message$1<"chat3.user.SubscribeUpdatesRequest"> & {
};

/**
 * Describes the message chat3.user.SubscribeUpdatesRequest.
 * Use `create(SubscribeUpdatesRequestSchema)` to create a new message.
 */
export declare const SubscribeUpdatesRequestSchema: GenMessage<SubscribeUpdatesRequest>;

/**
 * Обновление (update)
 *
 * @generated from message chat3.user.Update
 */
export declare type Update = Message$1<"chat3.user.Update"> & {
  /**
   * @generated from field: string update_id = 1;
   */
  updateId: string;

  /**
   * @generated from field: string tenant_id = 2;
   */
  tenantId: string;

  /**
   * @generated from field: string user_id = 3;
   */
  userId: string;

  /**
   * @generated from field: string entity_id = 4;
   */
  entityId: string;

  /**
   * @generated from field: string event_type = 5;
   */
  eventType: string;

  /**
   * @generated from field: google.protobuf.Struct data = 6;
   */
  data?: JsonObject;

  /**
   * @generated from field: double created_at = 7;
   */
  createdAt: number;
};